requirements.txt
fastapi[standard]
transformers
torch
uvicorn[standard]
python-dotenv